use senderName to identify the type of device
Also set the device instance-name in the signalk message to vedirectx, where-in x
is the last character of the service name. Fe. the 1 in
com.victronenergy.battery.ttyO1.

Thats only temporarily, something better will come later, and by then we'll also
drop vedirect from it, and replace it by something better.