add MQTT as an alternative to D-Bus
The benefit of this is being able to retrieve the data from another device. For
example running Venus OS on a Victron CCGX, and running signalk-node-server on
a raspberrypi.

Technically speaking the same can maybe/probably also be done with D-Bus, by
making that bind to a network interface. But thats another story.

This feature requires the MQTT service in Venus to be switched on. See
Settings -> Services -> MQTT.