feature: connect to D-Bus via tcp
By default, the system D-Bus on a Venus device is not listening on
tcp. To make it do so, add these three lines to /etc/dbus-1/system.conf
on the Venus device:

  <listen>tcp:host=0.0.0.0,port=78</listen>
  <auth>ANONYMOUS</auth>
  <allow_anonymous/>

Reboot afterwards. And also remember that any update of the Venus
device will override these (and any other) changes to the rootfs.
Perhaps someday this can be added to Venus as a real setting instead
of a hack.

Make sure to only do this on a trusted network.